pyproject.toml
[project]
name = "myportfolyoapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "cloudinary>=1.44.1",
    "colorlog>=6.10.1",
    "cryptography>=46.0.3",
    "discord-logger>=1.2.5",
    "django==6.0",
    "django-autoslug>=1.9.9",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-import-export>=4.4.0",
    "django-unfold>=0.76.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.4",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "markdown>=3.10",
    "numpy>=2.4.1",
    "pgvector>=0.4.2",
    "psycopg2>=2.9.11",
    "pyotp>=2.9.0",
    "sentry-sdk>=2.49.0",
    "unidecode>=1.4.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "lab>=8.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "tox>=4.34.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["api*", "config*"]
exclude = ["static*", "uploads*", "media*", "staticfiles*"]
